--- user_interface/src/components/ChatRoom.jsx
@@ -1,9 +1,0 @@
-import React, { Component } from "react";
-
-export class ChatRoom extends Component {
-  render() {
-    return <div></div>;
-  }
-}
-
-export default ChatRoom;
